config: support weighted URLs in To field (#5624) (#5625)
Backport #5624 to 0.29.

Currently in core we support weighted URLs like this in the config file:

```yaml
to:
  - https://a.example.com,1
  - https://b.example.com,2
```

However in the protobuf we use a separate `load_balancing_weights`
field:

```proto
message Route {
  repeated string to                     = 3;
  repeated uint32 load_balancing_weights = 37;
}
```

This PR updates the code to convert from protobuf so that it also
supports weights directly in the `to` addresses. The existing
`load_balancing_weights` behavior is preserved and will take precedence
when provided.


[ENG-2398](https://linear.app/pomerium/issue/ENG-2398/enterprise-api-upstream-weight-is-not-parsed)

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review