ci: update build job
Set 'fail-fast: false' so that if one of these jobs fails it will not
cancel the other. These have have unfortunately become very flaky, and
the current behavior is not helpful.

Remove 'go-version' from the matrix as it is currently unused and it
does not match the actual Go version in use.