Move validation logic into config module
The private validation logic defined in databroker was only used in
one place, and it was somewhat duplicating the logic provided in the
config module already. As the config module didn't have the cryptutil
validation in place, I have moved the defintion over to the config so
that it can be applied to any config usage throughout the code base.

This may have some unintended consequences if the validation is more
aggressive with the config module. In such case, we should split the
validation logic into multiple functions, so that at least the
consuming code (in this case databroker) should not need to know too
much about the implementation logic, and only rely on the specific
validation it needs.