ssh: always set non-nil args list for internal cli (#5740)
When accessing the internal CLI with no arguments, the cobra CLI
defaults its args to os.Argv, which are most likely going to be
`--config /some/config/file.yaml`, resulting in a cryptic error if the
internal CLI is accessed with no args (and the portal is disabled). This
defaults the internal CLI's args to a non-nil empty string list, so that
it does not default to argv.