databroker: move gc to public interface, abort sync calls with invalid record versions (#5753)
The databroker maintains a list of record changes over time. Those
changes are retrieved via the databroker `Sync` call. Both the in-memory
and postgres storage backends periodically remove old changes.

Since this logic is necessary for any storage backend, add a `Clean`
method to the public interface and start the background cleaner in the
databroker server rather than the storage backend itself. This makes it
easier to test.

Currently if a client attempts to `Sync` changes, but they've fallen too
far behind, we would not detect this and the client would never receive
the missing changes. The correct behavior is to abort the `Sync`, which
signals the client to do a full re-sync via `SyncLatest`. This PR
attempts to implement that behavior.

There is some risk with this change. If the assumptions I'm making about
how clients work turn out not to be true, we could end up with clients
continuously attempting to call `Sync` and never recovering.

-
[ENG-2640](https://linear.app/pomerium/issue/ENG-2640/core-databroker-sync-may-miss-changes)

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review