authenticate: cache authenticators (#5780)
## Summary
Currently we create a new authenticator on every request. When creating
sessions from incoming IdP tokens this results in HTTP requests to the
OIDC discovery document and JWKS keys every time. These are unnecessary
HTTP requests and should be cached.

Update the `defaultGetIdentityProvider` so that it will cache
authenticators and re-use them. We actually have very similar
functionality in the identity manager, but I wanted to minimize the
number of changes since this will need to be backported. Future
improvements would be to unify this into a single authenticator cache
and also handle frequent changes to IdP settings. (as of now they will
just stay cached until restart)

A separate access token validator is now being used. This lets us
properly trace the requests and adds a default 10 second timeout when
fetching keys. Because of the way the OIDC `RemoteKeySet` works, it
doesn't use the passed-in context, but rather uses a background context,
which will never be cancelled. As a result, if an HTTP request for keys
were to get stuck, we'd never be able to resolve. With a timeout we
should be able to recover. A future improvement would be to implement a
new custom `KeySet` that uses the passed-in context rather than a
background context for fetching keys. This would allow the spans to
occur in the proper trace.

## Related issues
-
[ENG-2701](https://linear.app/pomerium/issue/ENG-2701/authenticate-idp-access-token-verification-always-retrieves-the-openid)


## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review