config: add dns options (#5789)
Add new DNS options for customizing the behavior of DNS in Envoy.

- `dns_udp_max_queries`: sets the maximum number of UDP queries before a
new UDP port is used (default in envoy is unlimited, new default in
Pomerium will be 100)
- `dns_use_tcp`: uses TCP instead of UDP for DNS queries (default in
envoy is to use UDP)
- `dns_query_tries`: sets the number of retries for a dns lookup
(default in envoy is 4)
- `dns_query_timeout`: the timeout for the DNS query (default in envoy
is 5 seconds)

The existing `dns_lookup_family` option was moved along with these
settings to a new struct in the code. There should be no change in
behavior for it.

-
[ENG-2684](https://linear.app/pomerium/issue/ENG-2684/core-expose-dns-options)

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review