core: on-disk databroker storage (#5774)
## Summary
This adds a new `file` databroker storage provider that stores data
on-disk. The DSN format is: `file:///path/to/file`.

This storage provider is implemented with a pebble key-value store.
Pebble implements a global, sorted collection of key-value pairs. We
break up the databroker storage into various "key spaces". Each key
space is distinguished via a single-byte prefix. All the keys for that
key space will have the same prefix.

- `Lease`: used to implement distributed locking: key is the lease name,
value is the lease id and the expires at timestamp.
- `Metadata`: metadata about the store: server version and migration
version.
  - The server version is randomly generated for an empty database. 
  - Migrations are used for future changes to the on-disk format.
- `Options`: options for a record type: key is the record type, value is
the protobuf-encoded options.
- Options are currently only used to set a capacity limit for a record
type (for example for events).
- `RecordChange`: each record as they are updated: key is the version,
value is the protobuf-encoded record.
- Record changes are like a write-ahead log. They are used by clients
via the `Sync` call to receive changes as they are made.
- In addition to `Put`s, `Delete`s are also stored as record changes.
(tombstones)
- `RecordChangeIndexByType`: an index of record type to version: key is
the record type + the version, value is nil.
- This index allows us to implement `Sync` with a record type filter
efficiently.
- `Record`: the latest version of each record: key is the record type +
the record id, value is the protobuf-encoded record.
- `RecordIndexByTypeVersion`: an index of record type and version to
record id: key is the record type + the version, value is the record id.
  - This index is used to enforce capacity limits.
- `RegistryService`: registered services: key is kind + endpoint, value
is the expires at timestamp.
- We maintain an in-memory index by the expires at timestamp to delete
services when they are expired.

We use standard protobuf wire-encoding for protobuf types. Integers are
big-endian uint64s (to preserve order). Times are uint64 unix timestamps
in microseconds. Keys are either fixed lengths based on the type (for
example, if the first part of a key is an integer, we know its 8 bytes
long), or separated with a 0x00 byte. This means strings cannot contain
a null byte or the split logic will not work.

`SyncLatest` supports efficient filtering for `id` and `$index`. This is
used both for typical calls to `SyncLatest`, but also for calls to
`Query`. The most common type of query (used for databroker lookups when
evaluating policies) looks like this:

```json
{ 
  "$or": [
    { "id": "VALUE" },
    { "$index": "VALUE" }
  ] 
}
```

This filter will be evaluated by doing a `get` on the `Record` key
space, and an address lookup against an in-memory CIDR index (using the
same `bart` library we use for the in-memory store). The results will
then be merged together using a set-union. (since records are already
ordered we can do this more efficiently than with a hash table) We may
want to do further query optimization in the future, but this handles
the most common queries.

## Related issues
-
[ENG-2679](https://linear.app/pomerium/issue/ENG-2679/core-implement-on-disk-databroker-storage)


## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review