grpcutil: add client manager (#5811)
## Summary
Add a new `ClientManager` in `grpcutil`. 

Typically with gRPC you create a single client connection to a static
endpoint and re-use it throughout your application. It handles all the
lifecycle management for underlying connections.

If your network is dynamic rather than static, you'd typically use DNS
(or a similar mechanism) to switch to different servers, and still rely
on a single client connection.

However if we are establishing connections to endpoints that are defined
from configuration, those endpoints can change over time. We could
maintain a map of gRPC client connections for each endpoint, but with
frequent configuration changes this is a memory leak or requires careful
coordination to close connections that are no longer used. This can
happen if endpoints change (you are connecting to different
databrokers), or if settings change (we are using a new shared key).

So instead this PR introduces a "lazy" client that establishes gRPC
connections on-demand, and closes them when they are idle using a timer
that's reset on every request. Once the lazy client is no longer used,
it will be garbage collected, and the idle timer will close the
connection avoiding the memory leak.


## Checklist

- [ ] reference any related issues
- [x] updated unit tests
- [ ] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review