databroker: add erroring server, update imports (#5812)
## Summary
Add a new `erroringServer` that always returns errors for all databroker
and registry methods.

This server will be used by a `clusteredServer` that will use it if
configuration is invalid.

Also update imports to be consistent. We have multiple packages named
`databroker` which can be confusing. When the `internal/databroker` uses
`grpc/databroker` it should always use the `databrokerpb` name now.

## Related issues
-
[ENG-2823](https://linear.app/pomerium/issue/ENG-2823/databroker-add-erroring-server)


## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review