fix: health checks should reflect clustered databroker member health (#5834)
## Summary

When pomerium is started the startup, health and readiness checks will
now appropriately reflect the status of clustered-mode.

When members of the cluster are leaders , the default checks are used to
evaluate startup, health, and readiness.

When members of the cluster are followers, they must received sync
messages from the leader, otherwise startup will not succeed, and they
will be marked unhealthy/unready.

When members cannot join the cluster, due to incorrect configuration,
they are marked as unhealthy.

## Related issues


[ENG-2885](https://linear.app/pomerium/issue/ENG-2885/health-check-when-cluster-databroker-leader-is-inaccessible-health)

## User Explanation

Clustered mode member health is now reflected in health checks.

## Checklist

- [X] reference any related issues
- [X] updated unit tests
- [X] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [X] ready for review