migrate to filippo.io/csrf (#5853)
Migrate off of github.com/pomerium/csrf and onto filippo.io/csrf. This 
new package does not rely on tokens/cookies for CSRF protection, but
instead on the HTTP headers set by modern browsers. (See
https://words.filippo.io/csrf/ for more context.)

However, the header-based approach is incompatible with the OAuth2 login 
flow, which currently relies on the ability to set a CSRF token in the
'state' parameter. To handle this case, fork and adapt some of the logic
from github.com/pomerium/csrf around setting and validating a CSRF
cookie.