databroker: support default path for file backend (#5874)
## Summary
Add support for a default path if the DSN for the file storage backend
is the empty string, `file:` or `file://`. Also add support for when the
DSN is just a path (`/tmp/example` becomes `file:///tmp/example`).

Also update the way we determine the data directory so that on a mac it
returns a directory in `~/Library/Application Support`, and change the
fall back from `/tmp/pomerium/data` to `/var/tmp/pomerium/data`, which
is a little more persistent.

I also moved the autocert data to its own folder, which makes this a
breaking change.

## Related issues
-
[ENG-2959](https://linear.app/pomerium/issue/ENG-2959/core-add-support-for-a-default-file-path-for-the-on-disk-databroker)

## Checklist
- [x] reference any related issues
- [ ] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review