sessions: rename State to Handle (#5890)
## Summary
Currently we have a type `sessions.State` that represents the data we store in a JWT cookie for sessions. It contains the session id along with several other fields, but does not contain the entirety of the session data. This is confusing, and a better name for this object would be a `sessions.Handle`, since it's mostly just a reference to an actual session stored in the databroker.

This PR renames `sessions.State` to `sessions.Handle`. This is a sweeping change, but should not change any behavior.

## Related issues
- [ENG-3001](https://linear.app/pomerium/issue/ENG-3001/core-session-improvements)


## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`, `dependencies`, `ci`)
- [x] ready for review