testenv: refactor certificate log matching (#5897)
The current failure output from testenv.LogRecorder.Match() when
matching a TLS certificate is confusing, as it contains a bunch of
fields that are not actually being searched for. Let's make the output
of the failure message consistent with the match logic, by extracting
the special-case logic out of the match() method and onto the
testenv.Certificate struct.