controlplane: add debug server (#5891)
## Summary
Add a new setting `debug_address` that allows a user to expose the debug
listener similar to the `metrics_address`. Add an index page and a
`config_dump` endpoint.

The current behavior is to start the debug server on a random port only
accessible to the local host. This is still the default behavior.
Exposing the debug server is dangerous because there are sensitive
values in the config.

## Related issues
-
[ENG-3047](https://linear.app/pomerium/issue/ENG-3047/core-admin-panel)

## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review