identity: refactor provider registration
Add a generic type parameter to the RegisterAuthenticator and
AuthenticatorConstructor types so we can simplify the
RegisterAuthenticator() calls.

The only issue is that the oidc.New() method doesn't fit the pattern,
but if we rename the existing method to oidc.NewWithOptions() then we
can add an oidc.New() method that does fit.