ssh: improve internal cli output management (#5913)
This contains a few improvements to how stdout and stderr from the
internal cli are handled:
- Separate stdout and stderr internally; stderr messages are now sent
via `ChannelExtendedData` with `SSH2_EXTENDED_DATA_STDERR` data type.
- When the client requests a pty for the session, stderr output now
emulates newline translation processing which is disabled when the
client enters raw mode.
- Commands not marked as interactive will now work as normal when a pty
is requested. This would previously show an error to avoid displaying
text that would not be rendered correctly by the client.