core: add pre-commit script to bump component versions (#5928)
## Summary
With per-component versions we need to make sure to bump versions when
files change. For example if a new config option is added we should bump
the config component version from `v0.1.0` to `v0.2.0`.

This PR adds a script to do this automatically. It's not an exhaustive
check but should hopefully catch common use cases.

## Related issues
-
[ENG-3152](https://linear.app/pomerium/issue/ENG-3152/core-add-pre-commit-check-for-component-versioning)


## Checklist

- [x] reference any related issues
- [ ] updated unit tests
- [x] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review