envoy: store version to avoid importing files package (#5955)
## Summary
When we import the `files` package under `envoy` it requires the envoy
files to have been downloaded since they aren't stored in git. This
ended up making any package which imports pomerium also have to download
the files and be built with the `embed_pomerium` tag.

Instead we can store the version of envoy we download separately and
only import that, which should make it possible to import pomerium
without requiring the build tag.

## Checklist

- [ ] reference any related issues
- [ ] updated unit tests
- [ ] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review