storage: use pebble for inmemory store (#5959)
## Summary
Switch the in-memory storage backend to use the `file` backend with the
special `memory://` DSN. There should be no change in behavior with this
PR, but it results in less code to maintain.

## Related issues


## User Explanation

<!-- How would you explain this change to the user? If this
change doesn't create any user-facing changes, you can leave
this blank. If filled out, add the `docs` label -->

## Checklist

- [x] reference any related issues
- [x] updated unit tests
- [ ] add appropriate label (`enhancement`, `bug`, `breaking`,
`dependencies`, `ci`)
- [x] ready for review