ssh: policy index api + in-memory implementation (#5948)
This adds a new PolicyIndexer interface, which is responsible for
processing updates to configuration, session records, and active ssh
streams. All active streams can subscribe to receive incremental updates
as state changes.

This allows reverse tunnel routes to scale, since each reverse tunnel
connection does not need to independently compute its own set of
authorized upstream tunnel routes. The implementation can be swapped out
to support different environments.

An in-memory implementation is included, as well as a conformance test
suite that can be run against any implementation of the PolicyIndexer
interface to validate that it works as expected in a variety of
scenarios.

---------

Co-authored-by: Denis Mishin <dmishin@pomerium.com>