authorize: fix two bugs in upstream policy
Attempting to evaluate the upstream policy for a tunnel route without an
ssh_policy would result in an error. Handle this case gracefully (return
a result with both "allow: false" and "deny: false", to match the result
of regular policy evaluation).

Make sure to set the ReverseTunnel field to true for upstream policy
evaluation.

Add a unit test exercising both of these issues.