ssh: refactor in-memory policy indexer (#5979)
This refactors some of the internals of the in-memory policy indexer to
properly account for multiple streams authorized to the same session,
but with different auth requests. Within a single session, upstream
tunnel authorized routes are evaluated and cached separately for each
unique AuthRequest within that session. If multiple streams connect and
have the same parameters in the AuthRequest, the cached routes will be
shared as expected.

There is a limit to how many different sets of authorized routes will be
cached (currently 5). If more streams with unique AuthRequests connect,
the oldest cached entries will be evicted.