ssh: fix EDS endpoint updates (#5984)
Keep the presence of entries in clusterEndpoints consistent with the
presence of resources in the eds cache, so we can ensure there is always
a correct default (empty) load assignment for new clusters, including
those that are deleted and re-added. If endpoints are updated before the
delta request comes in, the presence of the cluster in clusterEndpoints
means we can assume the cached load assignment is valid.

Related: https://linear.app/pomerium/issue/ENG-3303